requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==5.0.5
elementpath==2.2.1
fastapi==0.63.0
filelock==3.0.12
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==2.10
inflect==5.3.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mysql-connector-python==8.0.24
nltk==3.6.2
numpy==1.20.2
packaging==20.9
pandas==1.2.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
sentence-transformers==1.0.4
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.11
starlette==0.13.6
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transformers==4.5.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
